requirements.txt
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
rest_framework_simplejwt==0.0.2
djangorestframework-simplejwt==5.2.2